Creando ruta para buscar por username, para ordenar por fecha o por rol

+++ src/main/java/com/henryleon/API/service/jpa/UsuarioService.java
@@ -1,11 +1,12 @@
 package com.henryleon.API.service.jpa;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import com.henryleon.API.entity.Rol;
 import com.henryleon.API.entity.Usuario;
 import com.henryleon.API.repository.UsuarioRepository;
 import com.henryleon.API.service.UsuarioServiceInterface;
@@ -27,9 +28,29 @@ public class UsuarioService implements UsuarioServiceInterface {
 	}
 
 	@Override
-	public void eliminar(int id) {
-		usuarioRepo.deleteById(id);
+	public String eliminar(int id) {
+		if(usuarioRepo.existsById(id)) {
+			usuarioRepo.deleteById(id);
+			return "Deleted";
+		}
+		return "Entity does not exists";
 	}
 	
+//	Metodos de ordenamiento
+	@Override
+	public List<Usuario> orderByDate() {
+		return usuarioRepo.findAll(Sort.by("fechaCreacion"));
+	}
+	
+
+	@Override
+	public List<Usuario> orderByRol() {
+		return usuarioRepo.findAll(Sort.by("rolId"));
+	}
+	
+	
+	public Usuario findByUsernameIs(String username) {
+		return usuarioRepo.findByUsernameIs(username);
+	}
 	
 }
